+++ CHANGE_LOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 # Chat Widget
 
-## [Unreleased]
+## [1.6.0] 2023-12-19
 
 ### Added
 - Enabled customizations for the start chat error pane by adding new `IStartChatErrorPaneProps` interface